requirements.txt
APScheduler==3.6.3
asgiref==3.4.1
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2021.5.30
dj-database-url==0.5.0
Django==3.2.7
gunicorn==20.1.0
psycopg2==2.9.1
python-telegram-bot==13.7
pytz==2021.1
six==1.16.0
sqlparse==0.4.2
tornado==6.1
tzlocal==3.0
whitenoise==5.3.0
